summary: Fix userSummary serialization for anons
"Standard" serialization requires at least one post to compile
a non-empty list of users, including the target user. And the client
needs to have the target user in `users` property of reply, otherwise
they'll show 404 instead of a proper user profile header.

This "custom" serialization was implemented for Timelines v2 once.
We cannot reuse it from there without big (and dangerous) refactoring,
so here we only implement a simplified edge case we need.