Rewrite and rename the Post's method that checks visibility of post
Method now has name 'isVisibleFor' and receives User object or null as argument (not userId). It returns true if the given user can view this post or false otherwise. It checks post privacy and bans of viewer and post author.

This method is used in 'postAccessRequired' middleware and in some controllers.

There was a bug in the CommentLikesController, it was not properly check access to the post if liker has relations with the post author. Now attempt to like/unlike comment under such posts will return 403 'You can not see this post' answer.