Send realtime event on group updated_at change
Groups updated_at field changes when the new post or comment creates in the group. In this case all group subscribers now receives the 'user:update' event with the 'updatedGroups' field that contains an array of serialized updated groups.

It should help to keep up-to-date the groups list in the right column. Closes #FreeFeed/issues/19 (it will need also client support)