Process the EXIF orientation of different parts of the image separate
Exif orientation is a complex beast. The "Orientation" tag can be presents in several EXIF sections: in IFD0 (the image itself) or in IFD1 (image thumbnail). Different libraries and even browsers can read thumbnail's Orientation as an image Orientation. So now we explicitly read 'IFD0:Orientation' and 'IFD1:Orientation' from the image. We rotate it according to IFD0's and remove all orientation tags before saving the image.